requirements.txt
thop
autotorch
torch-encoding
